feat: navigate to post details on reply tap on media full screen page
